add system prompts
